Closes #12
